+++ src/mappings/Common/TokenTypes/TrackSegment/TrackSegments.js
@@ -1,12 +1,21 @@
 //The key is based on rarity
+const monacaTrackDesc = "Home to the illustrious FORMULA 1 GRAND PRIX DE MONACO (“Monaco Grand Prix”),  the Circuit de Monaco is a track laid out on the city streets of Monte Carlo and La Condamine in the principality of Monaco, one of the most glamorous and prestigious locations in motorsport. The Monaco Grand Prix was established in 1929 and is considered one of the world’s most important automobile races.";
 let MonacoTrackSegments = {
     '0' : { 
         name: 'Grand Prix de Monaco',
-        description: 'Grand Prix de Monaco is an Apex Track Segment'
+        description: monacaTrackDesc
     },
     '1' : {
         name: 'Grand Prix de Monaco',
-        description: 'Grand Prix de Monaco is a Legendary Track Segment'
+        description: monacaTrackDesc
+    },
+    '2' : {
+        name: 'Grand Prix de Monaco',
+        description: monacaTrackDesc
+    },
+    '4' : {
+        name: 'Grand Prix de Monaco',
+        description: monacaTrackDesc
     }
 }